% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ck.R
\name{ck}
\alias{ck}
\title{ck}
\usage{
ck(NoL, gtcel, Gsh, Gsc, midDL, Timeresinsec)
}
\arguments{
\item{NoL}{number of storage layers}

\item{gtcel}{quantile in celerity distribution for overland flow}

\item{Gsh}{shape par gamma distribution for lower case lambda}

\item{Gsc}{scale par gamma distribution for lower case lambda}

\item{midDL}{mean distance (from distance distribution, dd)}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}
}
\description{
Compute the sub-surface celerity
}
\examples{
\dontrun{
ck()
}
}
\keyword{celerity}

