% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbSteps.R
\name{nbSteps}
\alias{nbSteps}
\title{nbSteps}
\usage{
nbSteps(maxL, speed, Timeresinsec)
}
\arguments{
\item{maxL}{max distance (from distance distribution, DD)}

\item{speed}{celerity}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}
}
\description{
calculates the delay-steps for each level
}
\examples{
\dontrun{
nbSteps()
}
}
\keyword{celerity}

